Refactor: Consolidate route handlers and improve module structure.
This change significantly refactors the routing layer of your application
to improve code organization and reduce file clutter.

Key changes:
- Introduced `src/utils/request_models.rs` for shared request-related
  structs like `PrettyQuery`.
- Created a new `src/routes/core_routes.rs` module. This module now
  contains the route handlers and router definitions previously spread
  across `get.rs`, `post.rs`, `put.rs`, `patch.rs`, `delete.rs`,
  `options.rs`, `status.rs`, and `anything.rs`.
- Updated `src/routes/mod.rs` to reflect the new module structure,
  exporting `core_routes` and removing references to the deleted files.
- Deleted the individual route files that were consolidated into
  `core_routes.rs`.
- Modified `src/main.rs` to merge the single `core_routes::router()`
  instead of numerous individual routers. The `healthz` and `delay`
  routes remain separate as before.

These changes aim to make your codebase easier to navigate, understand,
and maintain by centralizing related routing logic and reducing
boilerplate.
