+++ src/utils/mod.rs
@@ -3,3 +3,4 @@
 pub mod json_response;
 pub mod error_response;
 pub mod server_config;
+pub mod request_models;
