+++ src/utils/request_models.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize)]
+pub struct PrettyQuery {
+    pub pretty: Option<bool>,
+}
